+++ src/main/scala/millfork/node/Node.scala
@@ -176,7 +176,7 @@ case class IfStatement(condition: Expression, thenBranch: List[ExecutableStateme
   override def getAllExpressions: List[Expression] = condition :: (thenBranch ++ elseBranch).flatMap(_.getAllExpressions)
 }
 
-case class WhileStatement(condition: Expression, body: List[ExecutableStatement]) extends ExecutableStatement {
+case class WhileStatement(condition: Expression, body: List[ExecutableStatement], increment: List[ExecutableStatement], labels: Set[String] = Set("", "while")) extends ExecutableStatement {
   override def getAllExpressions: List[Expression] = condition :: body.flatMap(_.getAllExpressions)
 }
 
@@ -188,12 +188,20 @@ case class ForStatement(variable: String, start: Expression, end: Expression, di
   override def getAllExpressions: List[Expression] = VariableExpression(variable) :: start :: end :: body.flatMap(_.getAllExpressions)
 }
 
-case class DoWhileStatement(body: List[ExecutableStatement], condition: Expression) extends ExecutableStatement {
+case class DoWhileStatement(body: List[ExecutableStatement], increment: List[ExecutableStatement], condition: Expression, labels: Set[String] = Set("", "do")) extends ExecutableStatement {
   override def getAllExpressions: List[Expression] = condition :: body.flatMap(_.getAllExpressions)
 }
 
 case class BlockStatement(body: List[ExecutableStatement]) extends ExecutableStatement {
   override def getAllExpressions: List[Expression] = body.flatMap(_.getAllExpressions)
+}
+
+case class BreakStatement(label: String) extends ExecutableStatement {
+  override def getAllExpressions: List[Expression] = Nil
+}
+
+case class ContinueStatement(label: String) extends ExecutableStatement {
+  override def getAllExpressions: List[Expression] = Nil
 }
 
 object AssemblyStatement {